requirements.txt
pandas
numpy
scipy
lightgbm
optuna
joblib
python-dotenv
matplotlib
tqdm
scikit-learn

